Fix reviewbot & codedex issue

+++ frameworks/player_lite/player_control/sink/player_audio_sink.cpp
@@ -156,7 +156,7 @@ void AudioSink::GetStatus(AudioSinkStatus &status)
 void AudioSink::UpdateAudioPts(int64_t lastPts, int64_t &timestamp, OutputInfo &renderFrame)
 {
     if (renderFrame.timeStamp == -1) {
-        float sampleCnt = (renderFrame.buffers[0].length / attr_.audAttr.channel) / AUDIO_SAMPLE_WIDTH_BYTE;
+        float sampleCnt = ((float)renderFrame.buffers[0].length / attr_.audAttr.channel) / AUDIO_SAMPLE_WIDTH_BYTE;
         float duration = (sampleCnt / attr_.audAttr.sampleRate) * MS_SCALE;
         renderFrame.timeStamp = lastPts + duration;
     }
